chore: move tests to existing suite to avoid argv process weirdness in jenkins
